requirements.txt
absl-py==0.12.0
astunparse==1.6.3
autopep8==1.5.6
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
flake8==3.9.2
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
Pillow==8.2.0
pip==21.0.1
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyOpenSSL==20.0.1
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools==52.0.0.post20210125
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
